+++ spec/turtle_spec.rb
@@ -61,7 +61,7 @@ describe RDF::N3::Reader do
                  end
               end.should raise_error(RDF::ReaderError)
             end
-          rescue Spec::Expectations::ExpectationNotMetError => e
+          rescue RSpec::Expectations::ExpectationNotMetError => e
             pending() {  raise }
           end
         end